server/api/requirements.txt
Django==2.0.7
pytz==2018.5
psycopg2==2.7.5
djangorestframework==3.8.2
django-filter==1.1.0
